Add an end-to-end test for entity state subscriptions.
