+++ server/src/test/java/io/spine/server/projection/ProjectionRepositoryTest.java
@@ -41,8 +41,8 @@ import io.spine.server.projection.given.EntitySubscriberProjection;
 import io.spine.server.projection.given.ProjectionRepositoryTestEnv.GivenEventMessage;
 import io.spine.server.projection.given.ProjectionRepositoryTestEnv.NoOpTaskNamesRepository;
 import io.spine.server.projection.given.ProjectionRepositoryTestEnv.SensoryDeprivedProjectionRepository;
-import io.spine.server.projection.given.ProjectionRepositoryTestEnv.TestProjection;
 import io.spine.server.projection.given.ProjectionRepositoryTestEnv.TestProjectionRepository;
+import io.spine.server.projection.given.TestProjection;
 import io.spine.server.storage.RecordStorage;
 import io.spine.system.server.EntityHistoryId;
 import io.spine.system.server.EntityStateChanged;
