+++ server/src/test/java/io/spine/server/projection/given/ProjectionRepositoryTestEnv.java
@@ -20,14 +20,9 @@
 
 package io.spine.server.projection.given;
 
-import com.google.common.collect.HashMultimap;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Multimap;
-import com.google.protobuf.Message;
 import io.spine.core.EventContext;
 import io.spine.core.MessageEnvelope;
 import io.spine.core.Subscribe;
-import io.spine.server.entity.TestEntityWithStringColumn;
 import io.spine.server.projection.Projection;
 import io.spine.server.projection.ProjectionRepository;
 import io.spine.test.projection.Project;
@@ -42,8 +37,6 @@ import io.spine.test.projection.event.PrjProjectDeleted;
 import io.spine.test.projection.event.PrjProjectStarted;
 import io.spine.test.projection.event.PrjTaskAdded;
 import org.checkerframework.checker.nullness.qual.Nullable;
-
-import java.util.Set;
 
 import static io.spine.testing.TestValues.randomString;
 
@@ -87,7 +80,7 @@ public class ProjectionRepositoryTestEnv {
 
     /** Stub projection repository. */
     public static class TestProjectionRepository
-            extends ProjectionRepository<ProjectId, TestProjection, Project> {
+            extends TestProjection.Repository {
 
         private @Nullable MessageEnvelope lastErrorEnvelope;
         private @Nullable RuntimeException lastException;
@@ -120,103 +113,6 @@ public class ProjectionRepositoryTestEnv {
             extends ProjectionRepository<ProjectId, NoOpTaskNamesProjection, ProjectTaskNames> {
         public NoOpTaskNamesRepository() {
             super();
-        }
-    }
-
-    /** The projection stub used in tests. */
-    public static class TestProjection
-            extends Projection<ProjectId, Project, ProjectVBuilder>
-            implements TestEntityWithStringColumn {
-
-        /** The event message history we store for inspecting in delivery tests. */
-        private static final Multimap<ProjectId, Message> eventMessagesDelivered =
-                HashMultimap.create();
-
-        public TestProjection(ProjectId id) {
-            super(id);
-        }
-
-        public static boolean processed(Message eventMessage) {
-            boolean result = eventMessagesDelivered.containsValue(eventMessage);
-            return result;
-        }
-
-        /**
-         * Returns the IDs of projection instances, which processed the given message.
-         *
-         * <p>Empty set is returned if no instance processed the given message.
-         */
-        public static Set<ProjectId> whoProcessed(Message eventMessage) {
-            ImmutableSet.Builder<ProjectId> builder = ImmutableSet.builder();
-            for (ProjectId projectId : eventMessagesDelivered.keySet()) {
-                if (eventMessagesDelivered.get(projectId).contains(eventMessage)) {
-                    builder.add(projectId);
-                }
-            }
-            return builder.build();
-        }
-
-        public static void clearMessageDeliveryHistory() {
-            eventMessagesDelivered.clear();
-        }
-
-        private void keep(Message eventMessage) {
-            eventMessagesDelivered.put(getId(), eventMessage);
-        }
-
-        @Subscribe
-        public void on(PrjProjectCreated event) {
-            // Keep the event message for further inspection in tests.
-            keep(event);
-
-            Project newState = getState().toBuilder()
-                                         .setId(event.getProjectId())
-                                         .setStatus(Project.Status.CREATED)
-                                         .setName(event.getName())
-                                         .build();
-            getBuilder().mergeFrom(newState);
-        }
-
-        @Subscribe
-        public void on(PrjTaskAdded event) {
-            keep(event);
-            getBuilder().addTask(event.getTask());
-        }
-
-        /**
-         * Handles the {@link PrjProjectStarted} event.
-         *
-         * @param event
-         *         the event message
-         * @param ignored
-         *         this parameter is left to show that a projection subscriber
-         *         can have two parameters
-         */
-        @Subscribe
-        public void on(PrjProjectStarted event,
-                       @SuppressWarnings("UnusedParameters") EventContext ignored) {
-            keep(event);
-            Project newState = getState().toBuilder()
-                                         .setStatus(Project.Status.STARTED)
-                                         .build();
-            getBuilder().mergeFrom(newState);
-        }
-
-        @Subscribe
-        public void on(PrjProjectArchived event) {
-            keep(event);
-            setArchived(true);
-        }
-
-        @Subscribe
-        public void on(PrjProjectDeleted event) {
-            keep(event);
-            setDeleted(true);
-        }
-
-        @Override
-        public String getIdString() {
-            return getId().toString();
         }
     }
